+++ test/loco/core_test.clj
@@ -3,6 +3,14 @@
         loco.core
         loco.constraints))
 
+(defmacro test-constraint-model
+  ([docstring model solution-maps]
+   `(testing ~docstring
+      (is (= (set ~solution-maps)
+             (set (solutions ~model))))))
+  ([model solution-maps]
+   `(test-constraint-model nil ~model ~solution-maps)))
+
 (deftest basic-test
   (-> (solutions
         [($in :x 1 3)
@@ -237,3 +245,30 @@
          ($= :b ($reify ($< 0 :a)))])
     (= '({:a 5 :b 1}))
     is))
+
+(deftest knapsack-test
+  (test-constraint-model
+   "Super basic knapsack constraint"
+   [($knapsack [1 1] [1 1] [1 1] 2 2)]
+   '({}))
+  (test-constraint-model
+   "Basic knapsack constraint"
+   [($knapsack [2 3]
+               [6 5]
+               [:x :y]
+               :total-weight
+               :total-value)
+    ($in :x 0 10)
+    ($in :y 0 10)
+    ($in :total-weight 0 10)
+    ($in :total-value 0 10)]
+   '({:x 0 :y 0 :total-weight 0 :total-value 0}
+     {:x 1 :y 0 :total-weight 2 :total-value 6}
+     {:x 0 :y 1 :total-weight 3 :total-value 5}
+     {:x 0 :y 2 :total-weight 6 :total-value 10}))
+  (test-constraint-model
+   "Knapsack constraint with negatives"
+   [($in :x -100 100)
+    ($in :y -100 100)
+    ($knapsack [1 1] [10 20] [:x :y] 10 10)]
+   '({:x 19 :y -9})))